requirements.txt
discord.py @ git+git://github.com/Rapptz/discord.py@45d498c1b76deaf3b394d17ccf56112fa691d160
discord-components==2.1.0
Pillow==8.4.0